l04_rag/requirements.txt
google-adk
faiss-cpu
langchain
